Migrate KudosBox and related components
- Split Kudos.jsx into separate files for each component
- Tweaked things a little bit such that the kudos box is always
  displayed (for now) with the data for the current user -- this needs
  to be fixed subsequently
